% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_above}
\alias{alert_above}
\title{Test for values greater than a specific value}
\usage{
alert_above(extent_type = "all", extent_value = 1, rule_value, items = "[ALL]")
}
\arguments{
\item{extent_type}{"all", "any", "last", "consecutive"}

\item{extent_value}{lower limit of extent. e.g. \code{extent_type="any"} and \code{extent_value=5} means alert if there are 5 or more values that satisfy the condition, in any position}

\item{rule_value}{value to test against. e.g. \code{rule_value=1} means alert if value is greater than 1}

\item{items}{vector of values in item_col that the rule should be applied to. Or "[ALL]" to apply it to all items.}
}
\value{
An \code{alert_rule} object
}
\description{
Configure a rule to test the time series for the presence of values greater than a specific value. Tolerance can be adjusted using the \code{extent_type} and \code{extent_value} parameters, see Examples for details.
}
\examples{
# alert if all values are greater than 50
ars <- alert_rules(alert_below(extent_type = "all", rule_value = 50))

# alert if there are 10 or more values that are greater than 50
# or if the last 3 or more values are greater than 60
# or if 5 or more values in a row are greater than 40
ars <- alert_rules(
  alert_below(extent_type = "any", extent_value = 10, rule_value = 50),
  alert_below(extent_type = "last", extent_value = 3, rule_value = 60),
  alert_below(extent_type = "consecutive", extent_value = 5, rule_value = 40)
)
}
