% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_below}
\alias{alert_below}
\title{alert_below}
\usage{
alert_below(extent_type = "all", extent_value = 1, rule_value, items = "ALL")
}
\arguments{
\item{extent_type}{"all", "any", "last", "consecutive"}

\item{extent_value}{lower limit of extent. e.g. `extent_type="any"` and `extent_value=5` means alert if there are 5 or more values that satisfy the condition, in any position}

\item{rule_value}{value to test against. e.g. `rule_value=1` means alert if value is less than 1}

\item{items}{vector of values in item_col that the rule should be applied to. Or "ALL" to apply it to all items.}
}
\value{
An `alert_rule` object
}
\description{
Less than
}
