% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_custom}
\alias{alert_custom}
\title{Create a custom alert rule}
\usage{
alert_custom(short_name, description, function_call, items = "ALL")
}
\arguments{
\item{short_name}{short name to uniquely identify the rule. Avoid spaces and special characters.}

\item{description}{description of what the rule checks for}

\item{function_call}{expression to be evaluated per item. See details.}

\item{items}{vector of values in item_col that the rule should be applied to. Or "ALL" to apply it to all items.

The function_call expression is called using `eval()` within a `dplyr::summarise()` after grouping by the `item_col`.
Column names that can be used explicitly in the expression are: `value`, `item`, `timepoint`}
}
\value{
An `alert_rule` object
}
\description{
Create a custom alert rule
}
\examples{
alert_custom(
  short_name = "my_rule_combo",
  description = "Over 3 missing values and max value is > 10",
  function_call = quote(sum(is.na(value)) > 3 && max(value, na.rm = TRUE) > 10)
)

alert_custom(
  short_name = "my_rule_doubled",
  description = "Last value is over double the first value",
  function_call = quote(rev(value)[1] > 2*value[1])
)
}
