% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_difference_above_perc}
\alias{alert_difference_above_perc}
\title{Test for when there is a percentage increase in latest values}
\usage{
alert_difference_above_perc(
  current_period,
  previous_period,
  rule_value,
  items = "[ALL]"
)
}
\arguments{
\item{current_period}{vector containing positions from end of time series to use for comparison}

\item{previous_period}{vector containing positions from end of time series to use for comparison. Can overlap with \code{current_period} if desired.}

\item{rule_value}{value to test against. e.g. \code{rule_value=5} means alert if percentage increase is greater than 5}

\item{items}{vector of values in item_col that the rule should be applied to. Or "\link{ALL}" to apply it to all items.}
}
\value{
An \code{alert_rule} object
}
\description{
Check if latest values are greater than in a previous period, over a particular percentage
}
\details{
Based on the mean of values in the two periods.
}
\examples{
# alert if mean of last 3 values is over 20\% greater than mean of the previous 12 values
ars <- alert_rules(
  alert_difference_above_perc(current_period = 1:3, previous_period = 4:15, rule_value = 20)
)
}
