% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_equals}
\alias{alert_equals}
\title{alert_equals}
\usage{
alert_equals(extent_type = "all", extent_value = 1, rule_value, items = "ALL")
}
\arguments{
\item{extent_type}{"all", "any", "last"}

\item{extent_value}{lower limit of extent. e.g. `extent_type="any"` and `extent_value=5` means alert if there are 5 or more values that satisfy the condition, in any position}

\item{rule_value}{value to test against. e.g. `rule_value=0` means alert if value == 0}
}
\value{
An `alert_rule` object
}
\description{
alert_equals
}
