% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_missing}
\alias{alert_missing}
\title{Test for missing values}
\usage{
alert_missing(extent_type = "all", extent_value = 1, items = "[ALL]")
}
\arguments{
\item{extent_type}{"all", "any", "last", "consecutive"}

\item{extent_value}{lower limit of extent. e.g. `extent_type="any"` and `extent_value=5` means alert if there are 5 or more missing values in any position}

\item{items}{vector of values in item_col that the rule should be applied to. Or "[ALL]" to apply it to all items.}
}
\value{
An `alert_rule` object
}
\description{
Configure a rule to test the time series for the presence of NA values. Tolerance can be adjusted using the `extent_type` and `extent_value` parameters, see Examples for details.
}
\examples{
# alert if all values are NA
ars <- alert_rules(alert_missing(extent_type = "all"))

# alert if there are 10 or more missing values in total
# or if the last 3 or more values are missing
# or if 5 or more values in a row are missing
ars <- alert_rules(
  alert_missing(extent_type = "any", extent_value = 10),
  alert_missing(extent_type = "last", extent_value = 3),
  alert_missing(extent_type = "consecutive", extent_value = 5)
)
}
\seealso{
[alert_rules()], [alert_equals()]
}
