% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_rules}
\alias{alert_rules}
\title{Create set of alert rules}
\usage{
alert_rules(...)
}
\arguments{
\item{...}{alerts to apply to the time series}
}
\value{
An `alert_rules` object
}
\description{
Create set of alert rules
}
\examples{
# alert if any values are NA
# or if all values are zero
ars <- alert_rules(
  alert_missing(extent_type = "any", extent_value = 1),
  alert_equals(extent_type = "all", rule_value = 0)
)
}
