% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{arrange_items}
\alias{arrange_items}
\title{Arrange/sort a df based on a list of items}
\usage{
arrange_items(df, item_order = NULL)
}
\arguments{
\item{df}{df to arrange}

\item{item_order}{named list with names corresponding to columns in the supplied \code{df}. List members are
either \code{TRUE} for ascending order, or a character vector of values contained in the named
column for explicit ordering. If \code{item_order = NULL}, the original order will be kept. See Details.}
}
\value{
data frame
}
\description{
Arrange/sort a df based on a list of items
}
\section{Details}{
 For \code{item_order}, the names of the list members should correspond to the column
names in the \code{df}. Any names that don't match will be ignored. When multiple columns are
specified, they are sorted together, in the same order as the list. If a list item is \code{TRUE} then
that column is sorted in ascending order. If a list item is a character vector then that column
is sorted in the order of the vector first, with any remaining values included alphabetically at
the end
}

