% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{colspec}
\alias{colspec}
\title{Specify relevant columns in df}
\usage{
colspec(timepoint_col, item_col, value_col, group_col = NULL)
}
\arguments{
\item{group_col}{optional}
}
\description{
Specify relevant columns in df
}
