% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_rmd_tab_group}
\alias{construct_rmd_tab_group}
\title{Dynamically generate a group of tabs for an rmd chunk}
\usage{
construct_rmd_tab_group(
  tab_group_name,
  df,
  timepoint_col,
  item_col,
  value_col,
  tab_col,
  tab_order = NULL,
  history_type = "value",
  timepoint_limits = c(NA, NA),
  fill_with_zero = FALSE,
  item_order = NULL,
  item_label = "Item",
  summary_cols = c("last_value", "mean_last14"),
  history_style = "bar",
  sync_axis_range = FALSE
)
}
\arguments{
\item{sync_axis_range}{}
}
\description{
Chunk options must contain `results = 'asis'`.
Function writes directly to the chunk using side-effects
}
