% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_rmd_tab_item}
\alias{construct_rmd_tab_item}
\title{Dynamically generate a single tab for an rmd chunk}
\usage{
construct_rmd_tab_item(
  df,
  timepoint_col,
  item_col,
  value_col,
  outputspec,
  timepoint_limits = c(NA, NA),
  fill_with_zero = FALSE,
  item_order = TRUE,
  tab_name = NULL,
  tab_level = 1
)
}
\arguments{
\item{df}{Data frame containing time series in long format}

\item{timepoint_col}{Column to be used for x-axes}

\item{item_col}{Column containing categorical values identifying distinct time series}

\item{value_col}{Column containing the time series values which will be used for the y-axes.}

\item{outputspec}{Specification for display of tab contents}

\item{timepoint_limits}{Set start and end dates for time period to include. Defaults to min/max of timepoint_col}

\item{fill_with_zero}{Replace any missing or NA values with 0? Useful when value_col is a record count}

\item{item_order}{vector of values contained in item_col, for ordering the items in the table. Any values not mentioned are included alphabetically at the end. If NULL, the original order as given by unique(item_col) will be used.}

\item{tab_name}{Character string to appear on parent tab}

\item{tab_level}{Child level for tab. Value of 1 creates a tab with rmd level "##".}
}
\value{
(invisibly) the supplied df
}
\description{
Chunk options must contain `results = 'asis'`.
Function writes directly to the chunk using side-effects
}
