% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data frame containing multiple time series in long format}
\format{
\subsection{\code{example_data}}{

A data frame with 3,903 rows and 4 columns:
\itemize{
\item timepoint - Dates for the time series
\item item - Labels to identify the different time series
\item value - Values for the time series
\item tab - Labels to group related time series into tabs
}
}
}
\usage{
example_data
}
\description{
Simulated data to cover a range of different behaviours of time series
}
\keyword{datasets}
