% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_prescription_numbers}
\alias{example_prescription_numbers}
\title{Example data frame containing numbers of antibiotic prescriptions in long
format}
\format{
\subsection{\code{example_prescription_numbers}}{

A data frame with 6,570 rows and 4 columns:
\itemize{
\item PrescriptionDate - The date the prescriptions were written
\item Antibiotic - The name of the antibiotic prescribed
\item Spectrum - The spectrum of activity of the antibiotic.
This value is always the same for a particular antibiotic
\item NumberOfPrescriptions - The number of prescriptions written for this
antibiotic on this day
\item Location - The hospital site where the prescription was written
}
}
}
\usage{
example_prescription_numbers
}
\description{
Simulated data to demonstrate package usage
}
\keyword{datasets}
