% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{export_grouped_table_html}
\alias{export_grouped_table_html}
\title{Save a grouped table to an html file}
\usage{
export_grouped_table_html(
  df,
  colspec,
  outputspec = NULL,
  page_title = "tinduck grouped table",
  save_directory = ".",
  save_filename = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{colspec}{use colspec()}

\item{outputspec}{use outputspec()}

\item{page_title}{title}

\item{save_directory}{where to save the file to}

\item{save_filename}{name for the file}

\item{show_progress}{noisy or quiet}

\item{...}{further params for rmarkdown::render()}
}
\value{
path to saved file
}
\description{
Save a grouped table to an html file
}
