% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{export_table_html}
\alias{export_table_html}
\title{Save a single table to an html file}
\usage{
export_table_html(
  df,
  colspec,
  page_title = "tinduck single table",
  save_directory = ".",
  save_filename = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{colspec}{list}

\item{page_title}{title}

\item{save_directory}{where to save the file to}

\item{save_filename}{name for the file}

\item{show_progress}{noisy or quiet}

\item{...}{further params for rmarkdown::render()}
}
\value{
path to saved file
}
\description{
Save a single table to an html file
}
