% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{initialise_widgets}
\alias{initialise_widgets}
\title{Initialise HTML widgets}
\usage{
initialise_widgets(plot_type)
}
\description{
If the output is being constructed in 'asis' chunks, there must also be at least one standard chunk that
contains the relevant widgets, otherwise they won't render. dyGraph also needs to be rendered with appropriate plot_type
}
