% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantis-package.R
\docType{package}
\name{mantis-package}
\alias{mantis}
\alias{mantis-package}
\title{mantis: Multiple Time Series Scanner}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate interactive html reports that enable quick visual review of multiple related time series stored in a data frame. For static datasets, this can help to identify any temporal artefacts that may affect the validity of subsequent analyses. For live data feeds, regularly scheduled reports can help to pro-actively identify data feed problems or unexpected trends that may require action. The reports are self-contained and shareable without a web server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/mantis}
  \item \url{https://ropensci.github.io/mantis/}
  \item Report bugs at \url{https://github.com/ropensci/mantis/issues}
}

}
\author{
\strong{Maintainer}: T. Phuong Quan \email{phuongquan567@outlook.com} (\href{https://orcid.org/0000-0001-8566-1817}{ORCID})

Authors:
\itemize{
  \item T. Phuong Quan \email{phuongquan567@outlook.com} (\href{https://orcid.org/0000-0001-8566-1817}{ORCID})
}

Other contributors:
\itemize{
  \item  University of Oxford [copyright holder]
  \item  National Institute for Health Research (NIHR) [funder]
  \item Rodrigo Pires [reviewer]
  \item Margaret Siple [reviewer]
}

}
\keyword{internal}
