% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_html.R
\name{output_table_interactive}
\alias{output_table_interactive}
\title{Create html table from prepared data frame}
\usage{
output_table_interactive(
  table,
  item_label = "Item",
  plot_label = "History",
  summary_cols = c("max_value"),
  plot_type = "bar",
  sync_axis_range = FALSE,
  bordered = TRUE,
  ...
)
}
\arguments{
\item{table}{from prepare_table()}

\item{item_label}{Label for first column}

\item{summary_cols}{vector of which summary columns to include}

\item{plot_type}{"bar" or "line"}

\item{sync_axis_range}{sync all history graphs to same y-axis range}
}
\value{
html table
}
\description{
Create html table from prepared data frame
}
