% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{outputspec}
\alias{outputspec}
\title{Specify output options}
\usage{
outputspec(
  history_type = "value",
  history_style = "bar",
  item_label = "Item",
  history_label = "History",
  summary_cols = c("max_value"),
  sync_axis_range = FALSE
)
}
\arguments{
\item{sync_axis_range}{}
}
\description{
Specify output options
}
