% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{outputspec_static_heatmap}
\alias{outputspec_static_heatmap}
\title{Specify output options for a static report containing heatmaps}
\usage{
outputspec_static_heatmap(
  fill_colour = "blue",
  y_label = NULL,
  item_order = NULL
)
}
\arguments{
\item{fill_colour}{colour to use for the tiles}

\item{y_label}{string for y-axis label. Optional}

\item{item_order}{named list corresponding to \code{item_cols} columns for ordering the
items in the output. List values are either \code{TRUE} for ascending order, or a character vector
of values contained in the named column for explicit ordering. If \code{item_order = NULL}, the
original order will be kept. See Details.}
}
\value{
An \code{outputspec()} object
}
\description{
Specify output options for a static report containing heatmaps
}
\section{Details}{
 For \code{item_order}, the names of the list members should correspond to the
\code{item_cols} columns previously specified in the \code{inputspec}. Any names that don't match will be
ignored. When multiple columns are specified, they are sorted together, in the same order of
priority as the list. If a list member is \code{TRUE} then that column is sorted in ascending order.
If a list member is a character vector then that column is sorted in the order of the vector
first, with any remaining values included alphabetically at the end.
}

