% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{outputspec_static_multipanel}
\alias{outputspec_static_multipanel}
\title{Specify output options for a static report containing grids of plots.}
\usage{
outputspec_static_multipanel(sync_axis_range = FALSE, y_label = NULL)
}
\arguments{
\item{sync_axis_range}{Set the y-axis to be the same range for all the plots.
X-axes are always synced.}

\item{y_label}{string for y-axis label. Optional}
}
\value{
An `outputspec()` object
}
\description{
Currently only creates scatter plots
}
