% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_table}
\alias{prepare_table}
\title{Convert supplied df into required format for generating tables/plots}
\usage{
prepare_table(df, timepoint_col, item_col, value_col, history_type = "value")
}
\arguments{
\item{history_type}{"value" or "delta}
}
\value{
data frame
}
\description{
Supplied df needs to be long (at least for now)
}
