% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_custom}
\alias{alert_custom}
\title{Create a custom alert rule}
\usage{
alert_custom(short_name, description, function_call, items = NULL)
}
\arguments{
\item{short_name}{short name to uniquely identify the rule. Only include alphanumeric, '-', and
'_' characters.}

\item{description}{short description of what the rule checks for}

\item{function_call}{quoted expression containing the call to be evaluated per item, that returns either \code{TRUE} or \code{FALSE}.
Return value of \code{TRUE} means alert result is "FAIL". See Details.}

\item{items}{named list with names corresponding to members of \code{item_cols}. List members are
character vectors of values contained in the named column that the rule should be applied to.
If \code{items = NULL} the rule will be applied to all items. See Details.}
}
\value{
An \code{alert_rule} object
}
\description{
Specify a custom rule to run on the time series.
}
\section{Details}{


The supplied \code{function_call} is passed to \code{eval()} within a \code{dplyr::summarise()} after grouping
by the \code{item_cols} and ordering by the \code{timepoint_col}. Column names that can be used
explicitly in the expression are \code{value} and \code{timepoint}, and which refer to the values in the
\code{value_col} and \code{timepoint_col} columns of the data respectively. See Examples.

Use \code{items} to restrict the rule to be applied only to specified items. \code{items} can either be
NULL or a named list of character vectors. If \code{NULL}, the rule will be applied to all items. If
a named list, the names must match members of the \code{item_cols} parameter in the \code{inputspec}, (as
well as column names in the \code{df}), though can be a subset. If an \code{item_col} is not named in the
list, the rule will apply to all its members. If an \code{item_col} is named in the list, the rule
will only be applied when the \code{item_col}'s value is contained in the corresponding character
vector. When multiple \code{item_col}s are specified, the rule will be applied only to items that
satisfy all the conditions.
See Examples in \code{\link[=alert_rules]{alert_rules()}}
}

\examples{
ars <- alert_rules(
  alert_custom(
    short_name = "my_rule_combo",
    description = "Over 3 missing values and max value is > 10",
    function_call = quote(sum(is.na(value)) > 3 && max(value, na.rm = TRUE) > 10)
  ),
  alert_custom(
    short_name = "my_rule_doubled",
    description = "Last value is over double the first value",
    function_call = quote(rev(value)[1] > 2*value[1])
  )
)

}
\seealso{
\code{\link[=alert_rules]{alert_rules()}}, \code{\link[=alert_missing]{alert_missing()}}, \code{\link[=alert_equals]{alert_equals()}}, \code{\link[=alert_above]{alert_above()}}, \code{\link[=alert_below]{alert_below()}}, \code{\link[=alert_difference_above_perc]{alert_difference_above_perc()}}, \code{\link[=alert_difference_below_perc]{alert_difference_below_perc()}}
}
