% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_custom}
\alias{alert_custom}
\title{Create a custom alert rule}
\usage{
alert_custom(short_name, description, function_call, items = NULL)
}
\arguments{
\item{short_name}{short name to uniquely identify the rule. Only include alphanumeric, '-', and
'_' characters.}

\item{description}{short description of what the rule checks for}

\item{function_call}{call to be evaluated per item, that returns either \code{TRUE} or \code{FALSE}.
Return value of \code{TRUE} means alert result is "FAIL". See Details.}

\item{items}{named list with names corresponding to members of \code{item_cols}. List members are
character vectors of values contained in the named column that the rule should be applied to.
If \code{items = NULL} the rule will be applied to all items. See Details.}
}
\value{
An \code{alert_rule} object
}
\description{
The supplied \code{function_call} is passed to \code{eval()} within a \code{dplyr::summarise()} after grouping
by the \code{item_cols} and ordering by the \code{timepoint_col}. Column names that can be used explicitly
in the expression are: \code{value}, \code{item}, \code{timepoint}; and which refer to the values in the
\code{value_col}, \code{item_cols}, \code{timepoint_col} columns of the data respectively
}
\section{Details}{
 Use \code{items} to restrict the rule to be applied only to specified items.
\code{items} can either be NULL or a named list of character vectors. If \code{NULL}, the rule will be
applied to all items. If a named list, the names must match members of the \code{item_cols}
parameter in the \code{inputspec}, (as well as column names in the \code{df}), though can be a subset.
If an \code{item_col} is not named in the list, the rule will apply to all its members. If an
\code{item_col} is named in the list, the rule will only be applied when the \code{item_col}'s value is
contained in the corresponding character vector. When multiple \code{item_col}s are
specified, the rule will be applied only to items that satisfy all the conditions.
}

\examples{
ars <- alert_rules(
  alert_custom(
    short_name = "my_rule_combo",
    description = "Over 3 missing values and max value is > 10",
    function_call = quote(sum(is.na(value)) > 3 && max(value, na.rm = TRUE) > 10)
  ),
  alert_custom(
    short_name = "my_rule_doubled",
    description = "Last value is over double the first value",
    function_call = quote(rev(value)[1] > 2*value[1])
  )
)
}
