% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_difference_above_perc}
\alias{alert_difference_above_perc}
\title{Alert when there is a percentage increase in latest values}
\usage{
alert_difference_above_perc(
  current_period,
  previous_period,
  rule_value,
  items = "ALL"
)
}
\arguments{
\item{current_period}{vector containing positions from end of time series to use for comparison}

\item{previous_period}{vector containing positions from end of time series to use for comparison. Can overlap with `current_period` if desired.}

\item{rule_value}{value to test against. e.g. `rule_value=5` means alert if percentage change is greater than 5}

\item{items}{vector of values in item_col that the rule should be applied to. Or "ALL" to apply it to all items.}
}
\value{
An `alert_rule` object
}
\description{
Check if latest values are greater than in a previous period, over a particular percentage
}
\details{
Based on the mean of values in the two periods.
}
