% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_missing}
\alias{alert_missing}
\title{Test for missing values}
\usage{
alert_missing(extent_type = "all", extent_value = 1, items = NULL)
}
\arguments{
\item{extent_type}{"all", "any", "last", "consecutive"}

\item{extent_value}{lower limit of extent. e.g. \code{extent_type="any"} and \code{extent_value=5} means
alert if there are 5 or more missing values in any position}

\item{items}{named list with names corresponding to members of \code{item_cols}. List members are
character vectors of values contained in the named column that the rule should be applied to.
If \code{items = NULL} the rule will be applied to all items. See Details.}
}
\value{
An \code{alert_rule} object
}
\description{
Configure a rule to test the time series for the presence of NA values. Tolerance can be
adjusted using the \code{extent_type} and \code{extent_value} parameters, see Examples for details.
}
\section{Details}{
 Use \code{items} to restrict the rule to be applied only to specified items.
\code{items} can either be NULL or a named list of character vectors. If \code{NULL}, the rule will be
applied to all items. If a named list, the names must match members of the \code{item_cols}
parameter in the \code{inputspec}, (as well as column names in the \code{df}), though can be a subset.
If an \code{item_col} is not named in the list, the rule will apply to all its members. If an
\code{item_col} is named in the list, the rule will only be applied when the \code{item_col}'s value is
contained in the corresponding character vector. When multiple \code{item_col}s are
specified, the rule will be applied only to items that satisfy all the conditions.
See Examples in \code{\link[=alert_rules]{alert_rules()}}
}

\examples{
# alert if all values are NA
ars <- alert_rules(alert_missing(extent_type = "all"))

# alert if there are 10 or more missing values in total
# or if the last 3 or more values are missing
# or if 5 or more values in a row are missing
ars <- alert_rules(
  alert_missing(extent_type = "any", extent_value = 10),
  alert_missing(extent_type = "last", extent_value = 3),
  alert_missing(extent_type = "consecutive", extent_value = 5)
)

}
\seealso{
\code{\link[=alert_rules]{alert_rules()}}, \code{\link[=alert_equals]{alert_equals()}}, \code{\link[=alert_above]{alert_above()}}, \code{\link[=alert_below]{alert_below()}}, \code{\link[=alert_difference_above_perc]{alert_difference_above_perc()}}, \code{\link[=alert_difference_below_perc]{alert_difference_below_perc()}}, \code{\link[=alert_custom]{alert_custom()}}
}
