% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{alert_rules}
\alias{alert_rules}
\title{Create set of alert rules}
\usage{
alert_rules(...)
}
\arguments{
\item{...}{alerts to apply to the time series}
}
\value{
An `alert_rules` object
}
\description{
Create set of alert rules
}
