% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{bespoke_rmd_initialise_widgets}
\alias{bespoke_rmd_initialise_widgets}
\title{Initialise HTML widgets}
\usage{
bespoke_rmd_initialise_widgets(plot_type)
}
\arguments{
\item{plot_type}{"\code{bar}" or "\code{line}", depending on what will be used in real tables.}
}
\value{
A (mostly) invisible html widget
}
\description{
If the output is being constructed in 'asis' chunks, there must also be at least one standard chunk that
contains the relevant widgets, otherwise they won't render. dyGraph also needs to be rendered with appropriate plot_type
}
