% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{bespoke_rmd_tab_group}
\alias{bespoke_rmd_tab_group}
\title{Dynamically generate a group of tabs for an rmd chunk}
\usage{
bespoke_rmd_tab_group(
  df,
  inputspec,
  outputspec,
  alert_rules = NULL,
  timepoint_limits = c(NA, NA),
  fill_with_zero = FALSE,
  tab_order = NULL,
  tab_group_name = NULL,
  tab_group_level = 1
)
}
\arguments{
\item{df}{Data frame containing time series in long format}

\item{inputspec}{\code{\link[=inputspec]{inputspec()}} object specifying which columns in the supplied \code{df} represent the
"timepoint", "item", "value" and "tab" for the time series. A separate tab
will be created for each distinct value in the "tab" column.}

\item{outputspec}{\code{outputspec} object specifying the desired format of the html table(s). If
not supplied, default values will be used.}

\item{alert_rules}{\code{\link[=alert_rules]{alert_rules()}} object specifying conditions to test}

\item{timepoint_limits}{Set start and end dates for time period to include. Defaults to min/max of timepoint_col}

\item{fill_with_zero}{Logical. Replace any missing or NA values with 0? Useful when \code{value_col} is a record count}

\item{tab_order}{Optional vector containing values from tab_col in desired order of display}

\item{tab_group_name}{Character string to appear on parent tab label. If omitted or \code{NULL}, only the child tabs (and not the parent tab) will be created.}

\item{tab_group_level}{Integer specifying the nesting level of the parent tab. Value of 1 equates to rmd level "##". This is unaffected by the presence or not of \code{tab_group_name}.}
}
\value{
(invisibly) the supplied df
}
\description{
Chunk options must contain \code{results = 'asis'}.
Function writes directly to the chunk using side-effects
}
\details{
The supplied data frame should contain multiple time series in long format, i.e.:

\itemize{
\item one "timepoint" (date) column which will be used for the x-axes. This currently must be at a daily granularity, but values do not have to be consecutive.
\item one "item" (character) column containing categorical values identifying distinct time series.
\item one "value" (numeric) column containing the time series values which will be used for the y-axes.
\item Optionally, a "tab" (character) column containing categorical values which will be used to group the time series into different tabs on the report.
}

The \code{inputspec} parameter maps the data frame columns to the above.
}
