% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct_rmd_tab_group}
\alias{construct_rmd_tab_group}
\title{Dynamically generate a group of tabs for an rmd chunk}
\usage{
construct_rmd_tab_group(
  df,
  timepoint_col,
  item_col,
  value_col,
  tab_col,
  outputspec,
  timepoint_limits = c(NA, NA),
  fill_with_zero = FALSE,
  item_order = TRUE,
  tab_order = NULL,
  tab_group_name = NULL,
  tab_group_level = 1
)
}
\arguments{
\item{df}{Data frame containing time series in long format}

\item{timepoint_col}{Name of column to be used for x-axes}

\item{item_col}{Name of column containing categorical values identifying distinct time series}

\item{value_col}{Name of column containing the time series values which will be used for the y-axes.}

\item{tab_col}{Name of column containing categorical values which will be used to group the time series into different tabs.}

\item{outputspec}{Specification for display of tab contents}

\item{timepoint_limits}{Set start and end dates for time period to include. Defaults to min/max of timepoint_col}

\item{fill_with_zero}{Replace any missing or NA values with 0? Useful when value_col is a record count}

\item{item_order}{vector of values contained in item_col, for ordering the items in the table. Any values not mentioned are included alphabetically at the end. If NULL, the original order as given by unique(item_col) will be used.}

\item{tab_order}{Optional vector containing values from tab_col in desired order of display}

\item{tab_group_name}{Character string to appear on parent tab}

\item{tab_group_level}{integer specifying the nesting level of the parent tab. Value of 1 equates to rmd level "##".}
}
\value{
(invisibly) the supplied df
}
\description{
Chunk options must contain `results = 'asis'`.
Function writes directly to the chunk using side-effects
}
