% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{inputspec}
\alias{inputspec}
\title{Specify relevant columns in the source data frame}
\usage{
inputspec(timepoint_col, item_cols, value_col, tab_col = NULL, period = "day")
}
\arguments{
\item{timepoint_col}{String denoting the (date) column which will be used for the x-axes.}

\item{item_cols}{String denoting the (character) column containing categorical values identifying
distinct time series. Multiple columns that together identify a time series can be provided as a vector}

\item{value_col}{String denoting the (numeric) column containing the time series values which
will be used for the y-axes.}

\item{tab_col}{Optional. String denoting the (character) column containing categorical values
which will be used to group the time series into different tabs on the report.}

\item{period}{periodicity of the timepoint_col values. "day"/"month"/"year"}
}
\value{
A \code{inputspec()} object
}
\description{
Specify relevant columns in the source data frame
}
