% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert.R
\name{mantis_alerts}
\alias{mantis_alerts}
\title{Generate a data frame containing alert results}
\usage{
mantis_alerts(
  df,
  inputspec,
  alert_rules,
  filter_results = c("PASS", "FAIL", "NA"),
  timepoint_limits = c(NA, NA),
  fill_with_zero = FALSE
)
}
\arguments{
\item{df}{A data frame containing multiple time series in long format. See Details.}

\item{inputspec}{[`inputspec()`] object specifying which columns in the supplied `df` represent the
"timepoint", "item", "value"  and (optionally) "tab" for the time series. Each "item-tab" represents an individual time series}

\item{alert_rules}{[`alert_rules()`] object specifying conditions to test}

\item{filter_results}{only return rows where the alert result is in this vector of values}

\item{timepoint_limits}{Set start and end dates for time period to include. Defaults to min/max of timepoint_col. Can be either Date values or NAs.}

\item{fill_with_zero}{Replace any missing or NA values with 0? Useful when value_col is a record count}
}
\value{
tibble
}
\description{
Generate a data frame containing alert results
}
