% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{mantis_report}
\alias{mantis_report}
\title{Create an interactive time series report from a data frame}
\usage{
mantis_report(
  df,
  colspec,
  outputspec = NULL,
  report_title = "mantis report",
  dataset_description = "",
  save_directory = ".",
  save_filename = NULL,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{df}{A data frame containing multiple time series in long format. See Details.}

\item{colspec}{[`colspec()`] object specifying which columns in the supplied `df` represent the
"timepoint", "item", "value"  and (optionally) "tab" for the time series. If a "tab" column is specified, a separate tab
will be created for each distinct value in the column.}

\item{outputspec}{[`outputspec()`] object specifying the desired format of the html table(s). If
not supplied, default values will be used.}

\item{report_title}{Title to appear on the report.}

\item{dataset_description}{Short description of the dataset being shown.
This will appear on the report.}

\item{save_directory}{String specifying directory in which to save the report. Default is
current directory.}

\item{save_filename}{String specifying filename for the report, excluding any file extension. If
no filename is supplied, one will be automatically generated with the format
`mantis_report_YYMMDD_HHMMSS`.}

\item{show_progress}{Print progress to console. Default = `TRUE`.}

\item{...}{Further parameters to be passed to `rmarkdown::render()`. Cannot include any of
`input`, `output_dir`, `output_file`, `params`, `quiet`.}
}
\value{
A string containing the name and path of the saved report.
}
\description{
Accepts a data frame containing multiple time series in long format, generates a collection of
interactive time series plots for visual inspection, and saves the report to disk.
}
\details{
The supplied data frame should contain multiple time series in long format, i.e.:
* one "timepoint" (datetime) column which will be used for the x-axes. This currently must be at a daily granularity, but values do not have to be consecutive.
* one "item" (character) column containing categorical values identifying distinct time series.
* one "value" (numeric) column containing the time series values which will be used for the y-axes.
* Optionally, a "tab" (character) column containing categorical values which will be used to group the time series into different tabs on the report.
The `colspec` parameter maps the data frame columns to the above.
}
