% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{outputspec_interactive}
\alias{outputspec_interactive}
\title{Specify output options for the report}
\usage{
outputspec_interactive(
  plot_value_type = "value",
  plot_type = "bar",
  item_label = "Item",
  plot_label = "History",
  summary_cols = c("max_value"),
  sync_axis_range = FALSE,
  item_order = NULL,
  sort_by = NULL
)
}
\arguments{
\item{plot_value_type}{Display the raw "\code{value}" for the time series or display the calculated
"\code{delta}" between consecutive values.}

\item{plot_type}{Display the time series as a "\code{bar}" or "\code{line}" chart.}

\item{item_label}{String label to use for the "item" column in the report.}

\item{plot_label}{String label to use for the time series column in the report.}

\item{summary_cols}{Summary data to include as columns in the report. Options are
\code{c("max_value", "last_value", "last_value_nonmissing", "last_timepoint", "mean_value")}.}

\item{sync_axis_range}{Set the y-axis to be the same range for all time series in a table.
X-axes are always synced.}

\item{item_order}{vector of values contained in item_col, for explicitly ordering the items in
the table. Any values not mentioned are included alphabetically at the end. If TRUE, items
will be sorted in ascending order. If NULL, the original order as given by unique(item_col)
will be used.}

\item{sort_by}{column in output table to sort by. Can be one of \code{item}, \code{alert_overall}, or one
of the summary columns. Append a minus sign to sort in descending order e.g. \code{-max_value}.
Secondary ordering will be based on \code{item_order}.}
}
\value{
An \code{outputspec()} object
}
\description{
Specify output options for the report
}
