% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{outputspec_static_heatmap}
\alias{outputspec_static_heatmap}
\title{Specify output options for a static report containing heatmaps}
\usage{
outputspec_static_heatmap(fill_colour = "blue", y_label = NULL)
}
\arguments{
\item{fill_colour}{colour to use for the tiles}

\item{y_label}{string for y-axis label. Optional}
}
\value{
An `outputspec()` object
}
\description{
Specify output options for a static report containing heatmaps
}
