% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_table}
\alias{prepare_table}
\title{Convert supplied df into required format for generating tables/plots}
\usage{
prepare_table(
  df,
  timepoint_col,
  item_col,
  value_col,
  plot_value_type = "value",
  timepoint_limits = c(NA, NA),
  fill_with_zero = FALSE,
  item_order = NULL
)
}
\arguments{
\item{plot_value_type}{"value" or "delta"}

\item{timepoint_limits}{set start and end dates for time period to include. Defaults to min/max of timepoint_col}

\item{fill_with_zero}{replace any missing or NA values with 0? Useful when value_col is a record count}

\item{item_order}{vector of values contained in item_col, for ordering the items in the table. Any values not mentioned are included alphabetically at the end}
}
\value{
data frame
}
\description{
Supplied df needs to be long (at least for now)
}
